feat: refactor RPC server setup and configuration management
