+++ services/object-mapping-indexer/src/config.ts
@@ -1,3 +1,4 @@
+import 'dotenv/config'
 import { env, notNaN } from './utils/env.js'
 
 export const config = {
@@ -5,7 +6,7 @@ export const config = {
   requestSizeLimit: env('REQUEST_SIZE_LIMIT', '200mb'),
   corsAllowOrigins: env('CORS_ALLOW_ORIGINS', ''),
   nodeRpcUrl: env('NODE_RPC_URL'),
-  recoveryInterval: notNaN(Number(env('RECOVERY_INTERVAL', '100'))),
+  recoveryInterval: notNaN(Number(env('RECOVERY_INTERVAL', '1000'))),
   logLevel: env(
     'LOG_LEVEL',
     process.env.NODE_ENV === 'production' ? 'info' : 'debug',
